Add support for --ulimit command line option
This change adds support for --ulimit option to be passed directly to
'docker run'.

This is useful on Fedora running the build-env container as using 'yum
install' from inside the container may be very slow because the maximum
number of file descriptors reported by 'ulimit -n' is set to a crazy
high value. Passing '--ulimit nofile=1024' to run.py solves this issue.
This has been observed on Fedora, running CentOS 7.5 docker image, and
might be related to bug [1] which has been fixed in rpm v1.14.2.

Multiple --ulimit options can be specified and are all passed to docker.

If '--ulimit nofile=N' is not specified, a default nofile value of
1024 is passed to docker.

[1] https://bugzilla.redhat.com/show_bug.cgi?id=1537564

Signed-off-by: Thierry Escande <thierry.escande@vates.tech>