Explicitly require redhat-rpm-config
We had a xenserver-config-rpm RPM in our repositories, built from a SRPM
which is part of XenServer 8's build. It provides system-rpm-config, and
was pulled as a dependency of rpm-build instead of redhat-rpm-config when
we rebuilt the container.

But this RPM is not meant to be used yet in build environments, and was
breaking the build of the kernel RPM, for example.

So we explicitly require redhat-rpm-config.

Signed-off-by: Samuel Verschelde <stormi-xcp@ylix.fr>