Allow using podman as alternative to docker
podman does not interact with a daemon running as root, so unlike docker
it will not be able to modify files owned by arbitrary users including root.

To keep with current practices, we need to pass --userns=keep-id so it can
write into the user's directories: the `builder` user inside the container
is indeed equivalent to the user launching the container.

All other users including root are mapped UIDs not existing on the host
(typically 100000 and above), so any file created by such users in build
trees in the container will (by design) have those (numeric-only) owners
on the host if they remain after the build.

Signed-off-by: Yann Dirson <yann.dirson@vates.fr>