Force keep ncurses5 in base container
From XenServer 8.4 we inherit ncurses6, but linking against this one
could trigger issues if a variable must be used by both with different
types (ncurses project provides no such explicit information to help
detect such issues, though we might dig into ncurses-compat to find
out).

To be on the safe side, since any package rebuild must be able to link
against ncurses5, we have to prevent upgrade to ncurses6 until we know
we want it (since downgrading would be unnecessarily difficult), and
we have to pin openssh to an older version, since we did rebuild it
against ncurses6 aleady.

Signed-off-by: Yann Dirson <yann.dirson@vates.tech>