Set ghcr.io registry for the container
In some not-completely-understood cases, without an explicit registry
podman queries interactively for one.

Setting this explicitly before we actually publish images to github does
not seem to have any side-effect, even "forgetting the workaround for
linux/amd64/v2 for podman < 5.5.1" case is not as bad-looking as it would
with just a "localhost/" registry.

Signed-off-by: Yann Dirson <yann.dirson@vates.tech>