cli: change interface to use subcommands
This alineates usage of the tools with the trend from the last 5 years
or so, hopefully making it easier to use.

The CLI has been split into 3 actions: shell, run and build. They have
independent arguments, and they are visible when using --help.

It also requires users to specify the development environment or backend
that they want to use to complete the action. Currently only container
is supported.

Examples of commands:
- xcp-ng-dev container shell --rm --branch 8.3
- xcp-ng-dev container run --rm --branch 8.3 dune build @check
- xcp-ng-dev container build --rm --branch 8.3 ./xapi

Signed-off-by: Pau Ruiz Safont <pau.safont@vates.tech>