WIP provide a --local-repo flag
Set all local repos to priority 1, to avoid getting e.g. ocaml5 from
Alma.

FIXME: this curently assumes the "local repos" are plain dirs, and we
run createrepo_c on them.  This allows to use the output of "builddep"
as a cache without any other intermediate step, but that feels awkward
for the general case.  Maybe we should drop it?  Anyway the "builddep"
operation may change...