Add a version in the image tag
so we can change the interface used to communicate with the
container.
Using an image tag allows to automatically download the expected image on
the developer's computer in case of update.

In the future, if we distribute tagged versions of xcp-ng-dev (on pypi for
example), we might want to use that version instead.

Signed-off-by: Gaëtan Lehmann <gaetan.lehmann@vates.tech>