container build: free container space by default, replace --rm by --no-rm
Letting the user pass --rm like this is an ordinary container you may want
to keep around was a bad idea, as the script never reuses it.  So everyone
did want to use --rm and newcomers usually did not, until their disk got
full.

Keep the option not to destroy the container using --no-rm, but destroy it
by default.

Signed-off-by: Yann Dirson <yann.dirson@vates.tech>