init-container: rework how --no-exit works
Former behavior was "ignore all errors so we can always reach the point
where we launch the interactive shell".  This had the very bad side effect
that the script continues after first error, and can leave the container
in a state that does not help debugging.

Use an "exit trap" instead, which does not require the above-mentioned
hack.  Uses "| true" instead of "; exit 0" to neutralize any non-zero
exit code from the "shell" action, as that would not work any more - and
besides, it shows much better what we're doing.

Signed-off-by: Yann Dirson <yann.dirson@vates.tech>