HACK: avoid EPEL rsync repos
For some reason, for EPEL on ARM (and apparently no other cases)
dnf chooses to mostly use a RSYNC mirror, with rsync not even installed.
We should rather get this bug fixed, or be able to disable RSYNC
entries, but for now it is easier to jsut hardcode a french mirror.