Add support for --install
It can be useful to launch a given commands with a new packages
installed first.  It is very cumbersome to let "run" chain an install
command before the main command, so this would typically be encapsulated
in a script.

This allows to avoid this hassle with a new option, which can be
repeated to install several packages.