build.sh: disable overlay cache by default, add --overlay-cache
When the image gets built, the overlays resulting from `RUN dnf` lines
are cached.  When new packages are available (usually on remote
servers), those overlays do not get invalidated, and rebuilding the
image only gets us the old image from the overlay cache.

The official documentation for Dockerfile states that for invalidation
we can only use the --no-cache commandline option.

--overlay-cache enables the cache for use-cases needing it.

Signed-off-by: Yann Dirson <yann.dirson@vates.tech>